Write OneRegQuivIntActionFunction
This function is an attribute of a 1-regular quiver <quiver>. It is a binary function. Its first argument <x> is a vertex or arrow of <quiver>, its second is an integer <K>. This function then returns <x-k>, subject to the convention that vertices in the quiver look like
     ... i+1 --> i --> i-1 --> ...
and arrows look like
     ... --{a+1}--> o --{a}--> o --{a-1}--> ...
In other words, integers increase "against the grain" of the quiver.