Write RetractionOf2RegAugmentation
This is an attribute of a special biserial quiver, and a meaningful one for 2-regular augmentations (as constructed using the "2RegAugmentationOfQuiver" operation). In that case, this function takes a path of the augmentation and interprets it as a path in the original quiver. If we think of the inclusion map into the augmentation as a (based) section (with zero paths as base points) then this map is the corresponding (based) retraction. In remaining cases, this is a function that returns a warning message alongside a null response