fix overquiver lifting bug
Suppose the ground quiver <ground_quiv> of a SB algebra has 2-regular augmentation <2reg> and overquiver <oquiv>.

The contraction map goes from <oquiv> to <2reg>. In the component <LiftOf> of each generator of <oquiv>, we store the generator of <2reg> to which it contracts. If we subequently want to go from <2reg> to <ground_quiv>, then we will use the retraction (which is a separate function).