update <VertexIndexedSequenceFamilyOfQuiver>
In the present version of QPA, a quiver does not have a unique name. (The existing function <UniqueQuiverName> is really just a counter whose value increments.) Any attempt to incorporate "the unique name" of a quiver into the name of its vertex-indexed sequence family is therefore contrived at best and misguided at worst. I have removed such an attempt from a prior version of code